chores: update dependencies (#214)
* Upgrade dependencies and fix compilation errors

* Fix java-time package deprecation

* Update slf4j, sentry, exposed and kotlinx-cli to their latest versions

* Update autohelp to 1.4.0 which supports kord 0.8.0-M7

+++ src/main/kotlin/com/github/devcordde/devcordbot/commands/general/jdoodle/EvalCommand.kt
@@ -60,7 +60,7 @@ class EvalCommand : AbstractRootCommand() {
             int("language", "Die Sprache, in der das Codesnippet ist") {
                 required = true
                 Language.values().forEach {
-                    choice(it.humanReadable, it.ordinal)
+                    choice(it.humanReadable, it.ordinal.toLong())
                 }
             }
         }
